118. Protecting Resources with Authorization

+++ backend/routes/posts.js
@@ -65,10 +65,15 @@ router.put('/:id', checkAuth,multer({storage:storage}).single('image'),
         content: req.body.content,
         imagePath: imagePath
     })
-    console.log(post);
-    Post.updateOne({_id: req.params.id}, post).then((result) => {
-        console.log(result);
-        res.status(200).json({message:'Update successful'});
+    //console.log(post);
+    Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then((result) => {
+        //console.log(result);
+        if(result.modifiedCount > 0){
+            res.status(200).json({message:'Update successful'});
+        } else{
+            res.status(401).json({message:'Not Authorized'});
+
+        }
 
     })
 })
@@ -104,9 +109,15 @@ router.get('/:id', (req, res, next) => {
 });
 
 router.delete('/:id', checkAuth, (req, res, next) => {
-    Post.deleteOne({_id: req.params.id}).then(result => {
-        console.log(req.params.id);
-        res.status(200).json({message: 'Post deleted!'});
+    Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result => {
+        console.log(result);
+        // res.status(200).json({message: 'Post deleted!'});
+        if(result.deletedCount > 0){
+            res.status(200).json({message:'Deletion successful'});
+        } else{
+            res.status(401).json({message:'Not Authorized'});
+
+        }
     });
 });
